Add files via upload
A survey was conducted by Ben Silverstone for his Ph.D. thesis, with different colleges in the UK, to improve communication and create a more effective and efficient system for exchanging information. Participants were asked to complete a brief survey of 30 questions, which was kept confidential and anonymous, Out of 26 questions, no data was shared due to privacy concerns. The input of the people was greatly appreciated and was used to improve the organization's email communication practices. The poster displays the reports, and necessary actions can be carried out by the data-driven decision to improve the better exchange of information.